+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './lib';
 export * from './network';
 export * from './react';
 export * from './react-native';
+export * from './emotion';
