+++ src/react-native/component/Button/Button.type.ts
@@ -1,6 +1,6 @@
 import {TouchableOpacityProps} from 'react-native';
 
-import {theme} from '@/react/emotion/theme';
+import {theme} from '@/emotion/theme';
 
 export interface ButtonProps extends TouchableOpacityProps {
   themeColor?: keyof (typeof theme)['colors'];
