+++ src/react-native/index.ts
@@ -1,3 +1,1 @@
 export * from './component';
-
-export * from '../react/emotion/theme';
