+++ src/react/index.ts
@@ -1,3 +1,2 @@
 export * from './hook';
 export * from './provider';
-export * from './emotion';
